+++ src/cljs-lib/src/calva/read_config.cljs
@@ -4,7 +4,10 @@
             [vvvvalvalval.supdate.api :refer [supdate]]))
 
 (def ^:private updaters {:customREPLCommandSnippets [{:snippet str}]
-                         :customREPLHoverSnippets   [{:snippet str}]})
+                         :customREPLHoverSnippets   [{:snippet str}]
+                         :customPairForms           [identity]
+                         :customThreadingMacros     identity
+                         :aliasMap                  identity})
 
 (defn update-config-edn
   [config]